refact(win_errs): move HIDE_WINDOWS_KNOWN_ERRORS from main-code to test
